+++ resources/views/items/want_button.blade.php
@@ -1,7 +1,7 @@
-@if (Auth::user()->is_wanting($item->code))
+@if (Auth::check() && Auth::user()->is_wanting($item->code))
     {!! Form::open(['route' => 'item_user.dont_want', 'method' => 'delete']) !!}
         {!! Form::hidden('itemCode', $item->code) !!}
-        {!! Form::submit('Want', ['class' => 'btn btn-success']) !!}
+        {!! Form::submit('Wanted', ['class' => 'btn btn-success']) !!}
     {!! Form::close() !!}
 @else
     {!! Form::open(['route' => 'item_user.want']) !!}
